Fixed Conference, ReportThesis, Article and ScientificWork controllers, minor changes in UI

+++ ScientificReport/Controllers/ConferenceController.cs
@@ -1,8 +1,10 @@
 using System;
+using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ScientificReport.BLL.Interfaces;
+using ScientificReport.Controllers.Utils;
 using ScientificReport.DAL.Entities;
 using ScientificReport.DAL.Roles;
 using ScientificReport.DTO.Models.Conference;
@@ -13,21 +15,29 @@ namespace ScientificReport.Controllers
 	public class ConferenceController : Controller
 	{
 		private readonly IConferenceService _conferenceService;
+		private readonly IDepartmentService _departmentService;
+		private readonly IUserProfileService _userProfileService;
 
-		public ConferenceController(IConferenceService conferenceService)
+		public ConferenceController(
+			IConferenceService conferenceService,
+			IDepartmentService departmentService,
+			IUserProfileService userProfileService
+		)
 		{
 			_conferenceService = conferenceService;
+			_departmentService = departmentService;
+			_userProfileService = userProfileService;
 		}
 
 		// GET: Conference
 		public IActionResult Index(ConferenceIndexModel model)
 		{
-			model.Conferences = _conferenceService.GetPage(model.CurrentPage, model.PageSize);
-			model.Count = _conferenceService.GetCount();
+			model.Conferences = _conferenceService.GetPageByRole(model.CurrentPage, model.PageSize, User);
+			model.Count = _conferenceService.GetCountByRole(User);
 			return View(model);
 		}
 
-		// GET: Conference/Details/5
+		// GET: Conference/Details/{id}
 		public IActionResult Details(Guid? id)
 		{
 			if (id == null)
@@ -42,6 +52,11 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!UserHasPermission(conference))
+			{
+				return Forbid();
+			}
+
 			var conferenceDetails = new ConferenceDetails
 			{
 				Conference = conference
@@ -51,25 +66,23 @@ namespace ScientificReport.Controllers
 		}
 
 		// GET: Conference/Create
-		public IActionResult Create()
-		{
-			return View();
-		}
+		public IActionResult Create() => View();
 
 		// POST: Conference/Create
-		// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public IActionResult Create([Bind("Id,Topic,Date")] Conference conference)
 		{
-			if (!ModelState.IsValid) return View(conference);
+			if (!ModelState.IsValid)
+			{
+				return View(conference);
+			}
 
 			_conferenceService.CreateItem(conference);
 			return RedirectToAction(nameof(Index));
 		}
 
-		// GET: Conference/Edit/5
+		// GET: Conference/Edit/{id}
 		public IActionResult Edit(Guid? id)
 		{
 			if (id == null)
@@ -82,8 +95,12 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
-
 			
+			if (!UserHasPermission(conference))
+			{
+				return Forbid();
+			}
+
 			var conferenceEdit = new ConferenceEdit
 			{
 				Conference = conference
@@ -92,9 +109,7 @@ namespace ScientificReport.Controllers
 			return View(conferenceEdit);
 		}
 
-		// POST: Conference/Edit/5
-		// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+		// POST: Conference/Edit/{id}
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public IActionResult Edit(Guid id, ConferenceEdit conferenceEdit)
@@ -105,7 +120,15 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
-			if (!ModelState.IsValid) return View(conferenceEdit);
+			if (!UserHasPermission(conference))
+			{
+				return Forbid();
+			}
+			
+			if (!ModelState.IsValid)
+			{
+				return View(conferenceEdit);
+			}
 			try
 			{
 				_conferenceService.UpdateItem(conference);
@@ -122,7 +145,7 @@ namespace ScientificReport.Controllers
 			return RedirectToAction(nameof(Index));
 		}
 
-		// GET: Conference/Delete/5
+		// GET: Conference/Delete/{id}
 		public IActionResult Delete(Guid? id)
 		{
 			if (id == null)
@@ -135,17 +158,37 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(conference))
+			{
+				return Forbid();
+			}
 
 			return View(conference);
 		}
 
-		// POST: Conference/Delete/5
+		// POST: Conference/Delete/{id}
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public IActionResult DeleteConfirmed(Guid id)
 		{
+			if (!UserHasPermission(_conferenceService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			_conferenceService.DeleteById(id);
 			return RedirectToAction(nameof(Index));
+		}
+		
+		private bool UserHasPermission(Conference conference)
+		{
+			var user = _userProfileService.Get(User);
+			var department = _departmentService.Get(d => d.Staff.Contains(user));
+			return PageHelpers.IsAdmin(User) ||
+				   PageHelpers.IsHeadOfDepartment(User) &&
+				   _conferenceService.GetParticipators(conference.Id).Any(p => department.Staff.Contains(p)) ||
+				   _conferenceService.GetParticipators(conference.Id).Contains(user);
 		}
 	}
 }
